:recycle: move some folders
