+++ src/screens/components/event/noEvents/index.js
