+++ src/screens/components/talk/noTalks/index.js
