+++ src/screens/speaker/talk/submission/eventsSelection/eventsSelection.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { List, ListItem } from 'components/list'
 import Badge from 'components/badge'
 import IconLabel from 'components/iconLabel'
-import NoEvents from 'screens/speaker/components/noEvents'
+import NoEvents from 'screens/components/event/noEvents'
 import Status from 'screens/components/talk/status'
 import EventDates from 'screens/components/eventDates'
