+++ src/screens/speaker/talks/myTalks.jsx
@@ -8,7 +8,7 @@ import Titlebar from 'components/titlebar'
 import IconLabel from 'components/iconLabel'
 import Button from 'components/button'
 import RelativeDate from 'components/relativeDate'
-import NoTalks from 'screens/speaker/components/noTalks'
+import NoTalks from 'screens/components/talk/noTalks'
 
 import TalkInfo from './talkInfo'
