+++ src/screens/speaker/talks/talkInfo/index.js
